+++ zebra-consensus/src/block/tests.rs
@@ -2,6 +2,8 @@
 
 use super::*;
 
+use std::sync::Arc;
+
 use chrono::Utc;
 use color_eyre::eyre::{eyre, Report};
 use once_cell::sync::Lazy;
@@ -120,4 +122,21 @@ async fn check_transcripts() -> Result<(), Report> {
         transcript.check(block_verifier.clone()).await.unwrap();
     }
     Ok(())
+}
+
+#[test]
+fn time_check_past_block() {
+    // This block is also verified as part of the BlockVerifier service
+    // tests.
+    let block =
+        Arc::<Block>::zcash_deserialize(&zebra_test::vectors::BLOCK_MAINNET_415000_BYTES[..])
+            .expect("block should deserialize");
+    let now = Utc::now();
+
+    // This check is non-deterministic, but BLOCK_MAINNET_415000 is
+    // a long time in the past. So it's unlikely that the test machine
+    // will have a clock that's far enough in the past for the test to
+    // fail.
+    check::is_time_valid_at(&block.header, now)
+        .expect("the header time from a mainnet block should be valid");
 }
